# back-end/tsoa/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {},
		"securitySchemes": {}
	},
	"info": {
		"title": ""
	},
	"openapi": "3.0.0",
	"paths": {
		"/wands/get-all": {
			"get": {
				"operationId": "GetAllWand",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Wand"
				],
				"security": [],
				"parameters": []
			}
		},
		"/wands/create": {
			"post": {
				"operationId": "CreateWand",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Wand"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"shortDescription": {
										"type": "string"
									},
									"description": {
										"type": "string"
									},
									"image": {
										"type": "string"
									},
									"title": {
										"type": "string"
									}
								},
								"required": [
									"shortDescription",
									"description",
									"image",
									"title"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/wands/load-all": {
			"post": {
				"operationId": "LoadWands",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Wand"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			}
		},
		"/wands/update/{id}": {
			"put": {
				"operationId": "UpdateWand",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Wand"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"shortDescription": {
										"type": "string"
									},
									"description": {
										"type": "string"
									},
									"image": {
										"type": "string"
									},
									"title": {
										"type": "string"
									}
								},
								"required": [
									"shortDescription",
									"description",
									"image",
									"title"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/wands/delete/{id}": {
			"delete": {
				"operationId": "DeleteWand",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Wand"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/server-status": {
			"get": {
				"operationId": "GetServerStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"serverTime": {
											"type": "string"
										},
										"status": {
											"type": "string"
										}
									},
									"required": [
										"serverTime",
										"status"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Server Status"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/server-status/routes": {
			"get": {
				"operationId": "GetServerRoutes",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Server Status"
				],
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}